// models/city.model.ts
export type City = {
  id: number;
  name: string;
  ascii: string;
  alt_name: string;
  lat: number;
  long: number;
  feat_class: string;
  feat_code: string;
  country: string;
  cc2: string;
  admin1: string;
  admin2: string;
  admin3: string;
  admin4: string;
  population: number;
  elevation: number;
  dem: string;
  tz: string;
  modified_at: Date;
};